requirements.txt
# Core dependencies for GPT-4 MetaTrader5 Trading Bots
MetaTrader5>=5.0.4500
openai>=1.3.0
pandas>=1.5.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
PyYAML>=6.0
numpy>=1.24.0

# Optional dependencies for enhanced functionality
requests>=2.28.0
colorama>=0.4.6
tqdm>=4.65.0

# Development dependencies (uncomment if needed)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0